pyproject.toml
[project]
name = "llm-security"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "cachetools>=6.2.5",
    "colorlog>=6.10.1",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]

[project.scripts]
dev = "scripts.dev:main"
start = "scripts.start:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["api*", "core*", "security*", "utils*", "scripts*"]

[tool.uv]
package = true
